Refactor event functions
