+++ app/api/getAbsences/route.ts
@@ -1,67 +1,9 @@
-import { prisma } from '@utils/prisma';
 import { NextResponse } from 'next/server';
-
-export interface AbsenceWithRelations {
-  lessonDate: Date;
-  lessonPlan: string | null;
-  reasonOfAbsence: string;
-  notes: string | null;
-  absentTeacher: {
-    firstName: string;
-    lastName: string;
-    email: string;
-  };
-  substituteTeacher: {
-    firstName: string;
-    lastName: string;
-    email: string;
-  } | null;
-  location: {
-    name: string;
-    abbreviation: string;
-  };
-  subject: {
-    name: string;
-    abbreviation: string;
-  };
-}
+import { AbsenceWithRelations, getAbsencesFromDatabase } from './absences';
 
 export async function GET() {
   try {
-    const absences: AbsenceWithRelations[] = await prisma.absence.findMany({
-      select: {
-        lessonDate: true,
-        subject: {
-          select: {
-            name: true,
-            abbreviation: true,
-          },
-        },
-        lessonPlan: true,
-        reasonOfAbsence: true,
-        notes: true,
-        absentTeacher: {
-          select: {
-            firstName: true,
-            lastName: true,
-            email: true,
-          },
-        },
-        substituteTeacher: {
-          select: {
-            firstName: true,
-            lastName: true,
-            email: true,
-          },
-        },
-        location: {
-          select: {
-            name: true,
-            abbreviation: true,
-          },
-        },
-      },
-    });
+    const absences: AbsenceWithRelations[] = await getAbsencesFromDatabase();
 
     if (!absences.length) {
       return NextResponse.json({ events: [] }, { status: 200 });
